Hello World! Long time no see...
After months of playing CS, Project Rells is back and stronger than ever. It is worth mentioning that the Rells Team has now grown too.
So, we decided to move the old project to this new repository. This will still be maintained as to somewhat work with our new project and visions.

1. Server software: Changed the placeholder nickname for when separating the user and the password in the login packet upon a client joining the server.
2. Client software: Client software is now fully working. Made it so that you can use a new profile without it being previously saved.
3. Made a new library for making the process of developing and maintaining this CLI version easier.

	new file:   client.py
	new file:   rells_lib.py
	new file:   server.py